Splited DamageSystem in half and Renamed folder to "Damage"
