+++ Assets/Characters/Damage/Heddifs/Hediff.cs
