+++ Assets/Characters/Damage/Heddifs/InjuryHediff.cs.meta
