+++ Assets/Characters/Damage/HediffDefs.meta
