+++ Assets/Characters/Damage/HediffDefs/HediffDef.cs.meta
