+++ Assets/Characters/Damage/HediffDefs/InjuryHediffDef.cs
