+++ Assets/Characters/Damage/HediffSystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 39b180aff315c994cb1c8e681ec7529d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
