+++ Assets/Characters/Damage/HumanSkeleton.asset
